# uint128 ops are much faster without loops.